chore: Update TradeNFT and MyNFTs views
This commit updates the TradeNFT and MyNFTs views. In TradeNFT.vue, the `buyItem` method is modified to use the `view_nft.index` parameter instead of the `index` parameter. Additionally, the code related to displaying the owner's phone number is commented out. In MyNFTs.vue, the `ArgonSwitch` component is temporarily disabled and the loading spinner is hidden after a successful NFT edit.